requirements.txt
pandas
numpy
scikit-learn
selenium
beautifulsoup4
requests
nltk
vaderSentiment
spacy
matplotlib
seaborn
plotly
praw
pyspark
kaggle
Flask
python-dotenv
google-generativeai
